test5/requirements.txt
argon2-cffi>=23.1.0
pydantic==1.10
colorlog==6.8.2
fastapi==0.95.0
httpx>=0.24.0
psycopg2>=2.9.9
fastapi-authlib>=0.0.4
PyYAML>=6.0.1
sqlalchemy==1.4.46
PyJWT>=2.8.0
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.9
aiomqtt>=2.1.0